Status of app3-analysis on August 20, 2012.  Added data for two more examples.  Disabled saving .ascii and .TIFF files by default, now this can be done via command line options.

+++ fits/gaus1d.cpp
@@ -148,7 +148,7 @@ fit1dgaus_neldermead (gsl_vector * m, double *fit)
 	{
 	  printf
 	    ("%5d %10.3e %10.3e %10.3e %10.3e f() = %7.3e size = %.3f\n",
-	     iter, gsl_vector_get (s->x, 0), gsl_vector_get (s->x, 1),
+	     (int) iter, gsl_vector_get (s->x, 0), gsl_vector_get (s->x, 1),
 	     gsl_vector_get (s->x, 2), gsl_vector_get (s->x, 3), s->fval,
 	     size);
 	}
@@ -254,7 +254,7 @@ print_state_1d (size_t iter, gsl_multifit_fdfsolver * s)
     {
       printf ("iter: %3u x = % 15.8f % 15.8f % 15.8f % 15.8f  "
 	      "|f(x)| = %g\n",
-	      iter,
+	      (unsigned int) iter,
 	      gsl_vector_get (s->x, 0),
 	      gsl_vector_get (s->x, 1),
 	      gsl_vector_get (s->x, 2),
